don't explode on missing nested attrs
With flat attributes, if the specified attribute is not in the current
context, we omit the value entirely. It should work the same with nested
attributes.

If you want [:error, :message] and error will only be bound on one code
path, it should not throw an exception in the success case.