update core_components to 1.7.9 generated code
this was annoying because i've made a lot of changes to this file. i
ended up rebasing back to the initial generated code, pasting in the new
generated code, and then continuing the rebase to overlay all the
changes over the newer core components. then i reset back to origin/main
and pasted the updated core_components code in.

no idea if this was the beeest approach, but it was an approach.